refactor(tooling): split task catalog modules
